Extracted utility functions to utilityFunctions.kt
