+++ src/main/kotlin/com/manimdsl/runtime/ASTExecutor.kt
@@ -7,6 +7,7 @@ import com.manimdsl.frontend.*
 import com.manimdsl.linearrepresentation.*
 import com.manimdsl.shapes.Rectangle
 import com.manimdsl.stylesheet.Stylesheet
+import com.manimdsl.wrapCode
 import comcreat.manimdsl.linearrepresentation.*
 import java.util.*
 
@@ -45,36 +46,6 @@ class VirtualMachine(
                 displayLine.add(displayLine.lastOrNull() ?: 0)
             }
         }
-    }
-
-    fun wrapCode(code: MutableList<String>): MutableList<MutableList<String>> {
-        val wrappedCode = mutableListOf<MutableList<String>>()
-        for (line in code) {
-            val wrappedLine = mutableListOf<String>()
-            var index = 0
-            var prevIndex = 0
-            while (index < line.length) {
-                index = wrapLine(line.substring(index, line.length)) + prevIndex
-                wrappedLine.add(line.substring(prevIndex, index))
-                prevIndex = index
-            }
-            wrappedCode.add(wrappedLine)
-        }
-        return wrappedCode
-    }
-
-    fun wrapLine(line : String): Int{
-        val list = line.split(" ")
-        var counter = 0
-        var prevCounter = 0
-        for (word in list) {
-            counter += word.length + 1
-            if (counter >= WRAP_LINE_LENGTH) {
-                return prevCounter
-            }
-            prevCounter = counter
-        }
-        return line.length
     }
 
     fun runProgram(): Pair<ExitStatus, List<ManimInstr>> {
